% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armor_class.R
\name{armor.class}
\alias{armor.class}
\title{Armor Class Creator}
\usage{
armor.class(which.class, stats)
}
\arguments{
\item{which.class}{The class of the character will establish what armor will be issued to that character. It is deterministic with the assigned armor being from the equipment section of each class.}

\item{stats}{The calculated statistics used in stat.roll function. See it's help page for more.}
}
\value{
Will return a data frame of 2 with the first being the calcualted Armor Class and the second being the type of armor worn.
}
\description{
Issues a type of armor and calculates an armor class for a DnD character
}
\seealso{
Stats will come in a form produced by \code{\link{stat.roll}}
}
