% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_tables.R
\docType{data}
\name{character_tables}
\alias{character_tables}
\title{Class Tables}
\format{An object of class \code{list} of length 12.}
\source{
{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
character_tables
}
\description{
This contains the tables presented with each possible class presented as a list made up of the 12 playbale classes
}
\keyword{datasets}
