% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_roll.R
\name{check_roll.default}
\alias{check_roll.default}
\title{Check Rolls for Skills Default
Will allow for a skill check for a DnD character.}
\usage{
\method{check_roll}{default}(which.character, skill, ...)
}
\arguments{
\item{which.character}{The DnD character who is making the skill check}

\item{skill}{What skill is to be checked? Please no quotes.}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return a warning message about being passed a non-DnD object
}
\description{
Check Rolls for Skills Default
Will allow for a skill check for a DnD character.
}
\note{
There is no way to tell if the dice rolled used adv/dis after the fact so be careful in specifing the parameters.
}
