% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_roll.R
\name{check_roll}
\alias{check_roll}
\title{Check Rolls for Skills}
\usage{
check_roll(which.character, skill, ...)
}
\arguments{
\item{which.character}{The DnD character who is making the check}

\item{skill}{What skill is to be checked? Please no quotes.}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return the the check rolled including the proficiency bonus and character statistic modifier added properly.
}
\description{
Will allow for the a skill check for a DnD character
}
\note{
There is no way to tell if the dice rolled used adv/dis after the fact so be careful in specifing the parameters.
}
\examples{
Blacksmith_Jamestown <- DnD(Josh, 4)
check_roll(Blacksmith_Jamestown, History)
}
