% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Spells.R
\docType{data}
\name{Class_Spells}
\alias{Class_Spells}
\title{Spells Known by Class fully labeld}
\format{An object of class \code{list} of length 8.}
\source{
{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
Class_Spells
}
\description{
This is a list made up of the 8 spell casting classes. We do not preserve spell level but we do preserve the full name.
}
\keyword{datasets}
