% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_table.R
\docType{data}
\name{class_table}
\alias{class_table}
\title{Skills known by each Class}
\format{A data frame with 6 rows for the saving throws and 12 variables:
\describe{
  \item{Barbarian}{Class}
  \item{Bard}{Class}
  \item{Cleric}{Class}
  \item{Druid}{Class}
  \item{Fighter}{Class}
  \item{Monk}{Class}
  \item{Paladin}{Class}
  \item{Ranger}{Class}
  \item{Rogue}{Class}
  \item{Sorcerer}{Class}
  \item{Warlock}{Class}
  \item{Wizard}{Class}
  ...
}}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
class_table
}
\description{
This is a hashtable of the proficient saving throws for each class
}
\keyword{datasets}
