% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{damage.default}
\alias{damage.default}
\title{Current HP damage modifier}
\usage{
\method{damage}{default}(which.character, damage)
}
\arguments{
\item{which.character}{The DnD character who being hurt}

\item{damage}{The amount the character's health is to decrease by}
}
\value{
Will return the character with the new health status
}
\description{
Will allow for the DnD character's current HP to be automatically updated. No assignment is neccessary. This is a generic method with two functions, one for a DnD class object and one for a non-DnD class object.
}
\seealso{
\code{\link{damage.default}}
}
