% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{damage.DnD}
\alias{damage.DnD}
\title{Current HP damage modifier}
\usage{
\method{damage}{DnD}(which.character, damage)
}
\arguments{
\item{which.character}{The DnD character who being hurt}

\item{damage}{The amount the character's health is to decrease by}
}
\value{
Will return the character with the new health status
}
\description{
Will allow for the DnD character's current HP to be automatically updated. No assignment is neccessary. This is a generic method with two functions, one for a DnD class object and one for a non-DnD class object.
}
\examples{
DnD(Blacksmith_Jamestown, 4)
damage(Blacksmith_Jamestown, 20)
Blacksmith_Jamestown$hp
}
\seealso{
\code{\link{damage.default}}
}
