% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_background.R
\name{DnD}
\alias{DnD}
\title{DnD Character Generator}
\usage{
DnD(which.person = which.person, which.level = which.level, class = 0,
  race = 0, ...)
}
\arguments{
\item{which.person}{Who is this character? Any character string will work, don't need to use quotes.}

\item{which.level}{What level is the character? Please only choose numbers between 1 and 20.}

\item{class}{What class would you like the character to be? It will default to being randomly generated. Again, no quotes please.}

\item{race}{What race would you like the character to be? It will default to being randomly choosen. Again, no quotes please.}

\item{...}{The method used to roll the dice. Options include Commoner.Method (which is defaulted inside of DnD), Base.Method, Pilsfer.Method, Heroic.Method, and Ferguson.Method.}
}
\description{
Creates a DnD 5th edition character that is roughly built according to the rules. An important point is that the object is automatically built inside the function, DO NOT USE ASSINGMENT. This is redundant.
}
\details{
The following are all the pieces returned by the functon.
\itemize{
\item{Gamer:}{ which.person}
\item{Race:}{ The race of the character, one of 9 possible}
\item{Level:}{ which.level}
\item{Class:}{ The class of the character, one of 12 possible}
\item{Proficiency.Bonus:}{ The proficiency bonus modifier for known skills and saves}
\item{Saves:}{ Known saves of the class, 1 for known, 0 for unkown}
\item{Skills:}{ Known skills of the class}
\item{hp:}{ The maximum and current hit points of a character}
\item{Statistics:}{ The Stats and mod created for the character. Go to stat.roll help page for more indepth coverage. IMPORTANT: Stats are assigned according to the class inside the DnD function so do not try to pass this parameter in!}
\item{Weapons:}{ Three randomly assigned weapons regardless of proficiency or class}
\item{Armor.Class:}{ What armor is the character wearing and what is his AC?}
\item{Background:}{ Personality and motivation for the character}
\item{Spells:}{ The spells known/prepared for the character. See the help page for my.spells for more}
}
}
\note{
The characters produced here will have a class (as in R class) of "DnD"
}
\seealso{
Method and Stats are covered more in depth at \code{\link{stat.roll}}

Method for obtaining the spells for a character at \code{\link{my.spells}}

Method for obtaining the armor class for a character at \code{\link{armor.class}}

Method for obtaining the hit points for a character at \code{\link{hp.calculator}}
}
