% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DnD_info.R
\name{DnD_info}
\alias{DnD_info}
\title{A Dungeons and Dragons Weapon/Spell Information}
\usage{
DnD_info(name, type)
}
\arguments{
\item{name, }{The name of the weapon or spell}

\item{type, }{Either weapon or spell}
}
\description{
This function allows you to look up the stat blocks for selected weapons and spells in Dungeons and Dragons 5th Edition
This returns the webpage from DnDBeyond for the chosen weapon or spell
}
\examples{
DnD_info(longsword)

}
\keyword{DnD_info}
