% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DnD_roll.R
\name{DnD_roll}
\alias{DnD_roll}
\title{A Dungeons and Dragons dice rolling function}
\usage{
DnD_roll(n_dice, n_sides)
}
\arguments{
\item{n_dice, }{Number of Dice returned}

\item{n_sides, }{Number of Sides on each Dice returned}
}
\description{
This function allows you to choose how many dice you roll and how many sides the dice have
This returns the result of each roll and the sum of the rolls
}
\examples{
DnD_roll(4, 6)

}
\keyword{DnD_roll}
