% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heal.R
\name{heal.DnD}
\alias{heal.DnD}
\title{Heal a DnD Character}
\usage{
\method{heal}{DnD}(which.character, health)
}
\arguments{
\item{which.character}{The DnD character who is being healed.}

\item{health}{The amount the character should be healed.}
}
\value{
Will return the the simulated roll
}
\description{
Will allow for the DnD character to heal for some predetermined amount. No assignment is neccessary.
}
\seealso{
\code{\link{heal.DnD}}
}
