% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heal.R
\name{heal}
\alias{heal}
\title{Current HP health modifier}
\usage{
heal(which.character, health)
}
\arguments{
\item{which.character}{The DnD character who is being healed}

\item{health}{The amount the character's health is to increase by}
}
\value{
Will return the character with the new health status
}
\description{
Will allow for the DnD character's current HP to be automatically updated. No assignment is neccessary.
}
