% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp.R
\name{hp.calculator}
\alias{hp.calculator}
\title{Hit Points Calculator}
\usage{
hp.calculator(which.class = which.class, which.level = which.level)
}
\arguments{
\item{which.class}{The class of the character will establish what the hit dice will be issued to that character. It is deterministic with the assigned hit dice being from the section of each class in the player's hand book.}

\item{which.level}{The level of the character used to properly scale the character's health.}
}
\value{
Will return a data frame of 2 with the first being the calcualted Max Hit Points and the second being Current Hit Points.
}
\description{
Issues max and current hit points for a character based on level and class
}
\note{
The implementation of this function will allow only for half-round up method of calculating HP for each level above 1, i.e. we will not allow for a roll that occurs on the lower half of the dice but instead will move those to the lowest acceptable face.

The max and current hp produced will always be the same, this is simply to allow for easier use inside the DnD function.
}
