% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiative.R
\name{initiative.DnD}
\alias{initiative.DnD}
\title{Initiative Roller}
\usage{
\method{initiative}{DnD}(which.character, ...)
}
\arguments{
\item{which.character}{The DnD character who is making the initiative roll}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return the the simulated roll
}
\description{
Will allow for the DnD character to roll an intiative roll to be simulated. No assignment is neccessary.
}
\seealso{
\code{\link{initiative.DnD}}
}
