% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiative.R
\name{initiative}
\alias{initiative}
\title{Initiative Rolls}
\usage{
initiative(which.character)
}
\arguments{
\item{which.character}{The DnD character who is making the check}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return the the check rolled including the proficiency bonus and character statistic modifier added properly.
}
\description{
Will allow for the a initiative roll for a DnD character
}
\note{
There is no way to tell if the dice rolled used adv/dis after the fact so be careful in specifing the parameters.
}
