% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spells.R
\name{my.spells}
\alias{my.spells}
\title{Spell Creator}
\usage{
my.spells(which.level, which.class)
}
\arguments{
\item{which.level}{The level will determine how many spells of each level (if any) a character will know}

\item{which.class}{The class of the character will establish what spell list we need to work out of for this to generate properly. If a non-casting class is chosen a warning message will appear.}
}
\value{
Will return a list of 10 with spell levels 1-9 making up the first part and the tenth being Cantrips.
}
\description{
Create a spell list for a character based on class and level
}
\note{
Warlocks wil be given "Known Spells", "Cantrips", and "Spell Slots" given the uniquness of their class.

We are taking great liberties on known spells vs prepared spells inorder to stream line the process. More or less you will know and have prepared the exact number of spells you can cast at that level.
my.spells(10, "Warlock")
my.spells(20, "Bard")
my.spells(3, "Wizard")
my.spells(10, "Barbarian")
}
