% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_search.R
\name{para_search}
\alias{para_search}
\title{Finds the pattern matches in paragraphs of spells}
\usage{
para_search(pattern)
}
\arguments{
\item{pattern}{The pattern to be searched for}
}
\value{
Names of spells with paragraph matches
}
\description{
We searched the text of the spells for a matching character string and returned the names of the spells they were in.
}
