% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\docType{data}
\name{prof}
\alias{prof}
\title{Proficiency Vector}
\format{A vector of length 20}
\usage{
prof
}
\description{
This is a single vector of length 20 that gives the current proficiency bonus for a character
}
\keyword{datasets}
