% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_table.R
\docType{data}
\name{race_table}
\alias{race_table}
\title{Statistic Bonus given by each Race}
\format{A data frame with 6 rows and 9 variables:}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
race_table
}
\description{
This is a table of bonuses award to from each race to the six statistics
}
\keyword{datasets}
