% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest.default}
\alias{rest.default}
\title{Current HP maximizer}
\usage{
\method{rest}{default}(which.character)
}
\arguments{
\item{which.character}{The DnD character who is being healed}
}
\value{
Will return the character with the new health status
}
\description{
The default function for rest() in which a warning message will be returned if a non-DnD object is passed to it.
}
\seealso{
\code{\link{rest.DnD}}
}
