% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest.DnD}
\alias{rest.DnD}
\title{Current HP maximizer}
\usage{
\method{rest}{DnD}(which.character)
}
\arguments{
\item{which.character}{The DnD character who is being healed}
}
\value{
Will return the character with the new health status
}
\description{
Will allow for the DnD character's current HP to be automatically updated to the max HP the character has. No assignment is neccessary.
}
\seealso{
\code{\link{rest.default}}
}
