% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest.R
\name{rest}
\alias{rest}
\title{Current HP maximizer}
\usage{
rest(which.character)
}
\arguments{
\item{which.character}{The DnD character who is being healed}
}
\value{
Will return the character with the new health status
}
\description{
Will allow for the DnD character's current HP to be automatically updated to the max HP the character has. No assignment is neccessary. This is a generic method function with two subsidiary functions, one for DnD class objects and one for non-DnD class objects.
}
\seealso{
\code{\link{rest.DnD}}

\code{\link{rest.default}}
}
