% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling.R
\name{roll}
\alias{roll}
\title{Dice Rolling Function}
\usage{
roll(sides = 20, num.of.dice = 1, type = "reg")
}
\arguments{
\item{sides}{How many sides should be used to run the simulation? Defaults to 20}

\item{num.of.dice}{How many dice should be "rolled"? Defaults to 1}

\item{type}{Do you want to roll with advantage ("adv"), disadvantage ("dis"), or regular ("reg"). Defaults to regular.}
}
\value{
Will return the sum of the dice after taking into account adv/dis. There is also messages for a 20 or a 1 on a single 20 sided dice.
}
\description{
Allows for dice rolling simulation for any natural number of dice with any natural number of dice.
}
\note{
There is no way to tell if the dice rolled used adv/dis after the fact so be careful in specifing the parameters.
}
\examples{
roll()
roll(sides = 4)
roll(type = "adv")
}
