% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_throw.R
\name{saving_throw.default}
\alias{saving_throw.default}
\title{Saving Throws}
\usage{
\method{saving_throw}{default}(which.character, save, ...)
}
\arguments{
\item{which.character}{The DnD character who is making the saving throw}

\item{save}{What save is to be checked? Please no quotes.}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return the the saving throw including the proficiency bonus and character statistic modifier added properly.
}
\description{
Will allow for the a save throw for a DnD character. This *is* a generic function and will return a a warning message fro a non-DnD class object.
}
\note{
There is no way to tell if the dice rolled used adv/dis after the fact so be careful in specifing the parameters.
}
