% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skills_by_Class.R
\docType{data}
\name{Skills_by_Class}
\alias{Skills_by_Class}
\title{Skills known by each Class}
\format{A data frame with 18 rows and 12 variables:
\describe{
  \item{Columns}{The 12 possible classes, arranged alphabetical}
  \item{Rows}{The 18 skills, arranged in alphabetical ordering}
  ...
}}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
Skills_by_Class
}
\description{
This is a hashtable of the proficient skills for each class
}
\keyword{datasets}
