% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spells_by_Class.R
\docType{data}
\name{Spells_by_Class}
\alias{Spells_by_Class}
\title{Spells Known by Class}
\format{A data frame with 408 rows and 8 variables:
\describe{
  \item{Columns}{The 8 possible classes that know spells outside of pathological cases, arranged alphabetically}
  \item{Rows}{The 408 spells, arranged in alphabetical order}
  ...
}}
\source{
\url {Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
Spells_by_Class
}
\description{
This is a hashtable of the known spells for each class
}
\keyword{datasets}
