% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_table.R
\docType{data}
\name{stats_table}
\alias{stats_table}
\title{Prefered Statistics for Each Class}
\format{A data frame with 6 rows and 12 variables:
\describe{
  \item{Barbarian}{Class}
  \item{Bard}{Class}
  \item{Cleric}{Class}
  \item{Druid}{Class}
  \item{Fighter}{Class}
  \item{Monk}{Class}
  \item{Paladin}{Class}
  \item{Ranger}{Class}
  \item{Rogue}{Class}
  \item{Sorcerer}{Class}
  \item{Warlock}{Class}
  \item{Wizard}{Class}
  ...
}}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
stats_table
}
\description{
This is a data frame that was built using a general polling of regular players to provide the most prefered statistics for for each class. The discussion was held on April 5, 2017 on reddit's r/DnD page; whle not unanimous there was a general consensus found.
}
\keyword{datasets}
