% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stories.R
\docType{data}
\name{Stories}
\alias{Stories}
\title{Background information for characters}
\format{This is a list containing four peices of information:
\describe{
  \item{Personality}{151 unique personality traits}
  \item{Race}{9 playable races}
  \item{Class}{12 playable classes}
  \item{Background}{784 unique backgrounds to be given. Not neccessarly matching with Personality given above.}
  ...
}}
\source{
\url{whothefuckismydndcharacter.com}
}
\usage{
Stories
}
\description{
Contains everything neccessary to build a character story, race, class, and personality. Strung together with one chosen from each category, taken in order, will produce a sentence.
}
\keyword{datasets}
