% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weapons.R
\docType{data}
\name{weapons}
\alias{weapons}
\title{Weapons List}
\format{A vector of length 37:
\describe{
  \item{Element}{Character string with the name of a weapon inside DnD5e}
  
  ...
}}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
weapons
}
\description{
All weapons given, no differentiation between weapon types or characteristics (possibly a future update??).
}
\keyword{datasets}
