% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_table.R
\docType{data}
\name{class_table}
\alias{class_table}
\title{Skills known by each Class}
\format{A data frame with 6 rows and 12 variables:
\describe{
  \item{Columns}{The 12 possible classes, arranged alphabetical}
  \item{Rows}{The 6 saving throws, arranged in classic ordering (Str, Dex, Con, Int, Wis, Char)}
  ...
}}
\source{
\url{Dungeon and Dragons Fifth Edition Player's Hand Book}
}
\usage{
class_table
}
\description{
This is a hashtable of the proficient saving throws for each class
}
\keyword{datasets}
