% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiative.R
\name{initiative.default}
\alias{initiative.default}
\title{Initiative Roller}
\usage{
\method{initiative}{default}(which.character)
}
\arguments{
\item{which.character}{The DnD character who is making the initiative roll}

\item{...}{Additional arguments to be passed to the roll() function.}
}
\value{
Will return a warning message about being passed a DnD object.
}
\description{
Will allow for the DnD character to roll an intiative roll to be simulated. No assignment is neccessary.
}
\seealso{
\code{\link{initiative.default}}
}
